Break out tool abstraction
