+++ docker/user_tools/README.md
@@ -1,22 +1,58 @@
-# User Tools Directory
+# User Configs Directory
 
 ## Overview
-This directory is mounted inside the R2R Docker container and is intended for custom tools definitions. Any files placed here will be accessible to the application running in the container.
+This directory is mounted inside the R2R Docker container and is intended for custom tool files. Any files placed here will be accessible to the application running in the container.
 
 ## Usage
-1. Place your custom tool definitions in this directory.
-2. Update the `R2R_CONFIG_PATH` in the `r2r.env` or `r2r-full.env` files.
-3. The path format inside the container is: `/app/user_tools/<tool>.py`
+1. Place your custom tool definitions in this directory. Utilize the template structure demonstrated here.
+2. Include the tool in your your agent configuration.
 
-## Configuration
-The application uses the environment variable you set to locate your configuration file:
-```
-R2R_CONFIG_PATH=/app/user_configs/<config>.toml
-```
+## Creating a tool
+```python
+from core.base.agent.tools.base import Tool
 
-If you want to use a different filename, update the `R2R_CONFIG_PATH` variable in your environment file to point to your custom file, for example:
-```
-R2R_CONFIG_PATH=/app/user_configs/my_custom_config.toml
+
+class ToolNameTool(Tool):
+    """
+    A user defined tool.
+    """
+
+    def __init__(self):
+        super().__init__(
+            name="tool_name",
+            description="A natural language tool description that is shown to the agent.",
+            parameters={
+                "type": "object",
+                "properties": {
+                    "input_parameter": {
+                        "type": "string",
+                        "description": "Define any input parameters by their name and type",
+                    },
+                },
+                "required": ["input_parameter"],
+            },
+            results_function=self.execute,
+            llm_format_function=None,
+        )
+
+    async def execute(self, input_parameter: str, *args, **kwargs):
+        """
+        Implementation of the tool.
+        """
+
+        # Any custom tool logic can go here
+
+        output_response = some_method(input_parameter)
+
+        result = AggregateSearchResult(
+            generic_tool_result=[web_response],
+        )
+
+        # Add to results collector if context is provided
+        if context and hasattr(context, "search_results_collector"):
+            context.search_results_collector.add_aggregate_result(result)
+
+        return result
 ```
 
 ## Troubleshooting
