+++ py/core/__init__.py
@@ -82,8 +82,6 @@ __all__ = [
     "AgentConfig",
     "Conversation",
     "Message",
-    "Tool",
-    "ToolResult",
     "TokenResponse",
     "User",
     "AppConfig",
