+++ py/core/base/agent/__init__.py
@@ -3,8 +3,6 @@ from .agent import (  # type: ignore
     Agent,
     AgentConfig,
     Conversation,
-    Tool,
-    ToolResult,
 )
 
 __all__ = [
@@ -12,6 +10,4 @@ __all__ = [
     "Agent",
     "AgentConfig",
     "Conversation",
-    "Tool",
-    "ToolResult",
 ]
