+++ py/core/base/agent/tools/built_in/get_file_content.py
@@ -2,7 +2,7 @@ import logging
 from typing import Any, Optional
 from uuid import UUID
 
-from core.base.agent.tools.base import Tool
+from shared.abstractions.tool import Tool
 
 logger = logging.getLogger(__name__)
