+++ py/core/base/agent/tools/registry.py
@@ -6,7 +6,7 @@ import pkgutil
 import sys
 from typing import Callable, Optional, Type
 
-from .base import Tool
+from shared.abstractions.tool import Tool
 
 logger = logging.getLogger(__name__)
 
@@ -25,12 +25,7 @@ class ToolRegistry:
         self.built_in_path = built_in_path or os.path.join(
             os.path.dirname(os.path.abspath(__file__)), "built_in"
         )
-        self.user_tools_path = user_tools_path or os.path.join(
-            os.path.dirname(
-                os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-            ),
-            "user_tools",
-        )
+        self.user_tools_path = user_tools_path or "../docker/user_tools"
 
         # Tool storage
         self._built_in_tools: dict[str, Type[Tool]] = {}
@@ -131,7 +126,7 @@ class ToolRegistry:
                         try:
                             tool_instance = obj()
                             self._user_tools[tool_instance.name] = obj
-                            logger.info(
+                            logger.debug(
                                 f"Loaded user tool: {tool_instance.name}"
                             )
                         except Exception as e:
